% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autothreshold.value.func.ccsdf.R
\name{autothreshold.value}
\alias{autothreshold.value}
\title{Perform automatic thresholding on raster values.}
\usage{
autothreshold.value(raster, max.index, min.index, method)
}
\arguments{
\item{raster}{A `Raster` object to which thresholding will be applied.}

\item{max.index}{A numeric value specifying the maximum index for the 
thresholding.}

\item{min.index}{A numeric value specifying the minimum index for the 
thresholding.}

\item{method}{A string specifying the thresholding method to apply. 
See the documentation of the `autothresholdr::auto_thresh_apply_mask` 
function for supported methods.}
}
\value{
A numeric value representing the threshold identified for the raster 
data.
}
\description{
This function applies automatic thresholding to a raster object by first
binning its values into 256 levels based on the specified range, 
then applying the selected thresholding method to identify an appropriate 
threshold value. It is used for identifying a threshold value in
raster-based index data.
}
\keyword{internal}
