% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{calcs}
\alias{calcs}
\title{Perform Image Analysis and Index Calculation}
\usage{
calcs(
  photo,
  area,
  obs.areas,
  vis.files,
  nir.files,
  pic.format,
  manual.mask.test,
  mask.files,
  summary.file,
  chart,
  total.samples,
  index.,
  descriptors.,
  calculate.thresh,
  threshold.vector,
  descrip,
  threshold.method,
  start.time,
  chart.vals
)
}
\arguments{
\item{photo}{Integer. The index of the current photo being processed.}

\item{area}{Integer. The index of the area associated with the current photo.}

\item{obs.areas}{list of SpatialPolygons. 
Spatial polygons that set the areas where the samples are located in the 
pictures.}

\item{vis.files}{Character vector. File paths for visible spectrum images.}

\item{nir.files}{Character vector. File paths for near-infrared images.}

\item{pic.format}{Character. Image format, either "tif" or "jpg".}

\item{manual.mask.test}{Logical. Whether manual masking is used. If TRUE, 
provide mask.files argument}

\item{mask.files}{Character vector. File paths for manually created masks.}

\item{summary.file}{Character. Path to the CSV file where summary results are stored.}

\item{chart}{SpatialPolygons. Provided by chart.2 function that set position of color 
chart tiles in the pictures. With as many features as color tiles use for calibration.}

\item{total.samples}{Integer. Total number of samples to be processed.}

\item{index.}{Character vector. Names of the indices to be computed.}

\item{descriptors.}{Character vector. Statistical descriptors to compute.}

\item{calculate.thresh}{Logical. Whether threshold values should be computed.}

\item{threshold.vector}{Numeric vector. Predefined threshold values.}

\item{descrip}{Logical. Whether to calculate additional statistical descriptors.}

\item{threshold.method}{Character. The method used for threshold calculation.}

\item{start.time}{POSIXct. The start time of processing for logging execution time.}

\item{chart.vals}{Data frame. Calibration values for color correction.}
}
\value{
A list containing:
\item{data.frames}{A list of data frames with index values and classification results.}
}
\description{
calcs function processes a given image dataset, extracts relevant 
color and spectral information, applies thresholding techniques, and 
computes various index values for analysis.
}
\details{
The function executes the following steps:
- Reads input images and checks validity.
- Extracts color data and applies calibration.
- Computes selected indices and threshold values.
- Compares thresholding results with manual masks (if applicable).
- Saves results to the summary file.
}
\note{
This function is designed for internal use within an image processing pipeline.
}
\seealso{
\code{\link{raster.tiff.ccspectral}}, \code{\link{index.calc.fun}},
\code{\link{calculate.raster.thresh.fun}}
}
