% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.raster.thresh.fun.ccsdf.R
\name{calculate.raster.thresh.fun}
\alias{calculate.raster.thresh.fun}
\title{Calculate Raster Thresholds}
\usage{
calculate.raster.thresh.fun(
  list.raster.results,
  calculate.thresh,
  threshold.vector,
  threshold.method
)
}
\arguments{
\item{list.raster.results}{A list containing raster objects for different 
indices.}

\item{calculate.thresh}{A logical value. 
If TRUE, the function will calculate threshold values. 
If FALSE, it will use the provided threshold vector.}

\item{threshold.vector}{A numeric vector containing the threshold values for 
each raster index (used only when calculate.thresh is FALSE).}

\item{threshold.method}{The thresholding method to use for automatic
threshold calculation.}
}
\value{
A list containing two elements:
- A list of raster index cut values, where each element is a 
binary raster.
- A vector of threshold values used for the corresponding raster indices.
}
\description{
This function calculates threshold values for various raster indices using 
an automatic thresholding method.
It can process a list of raster results and apply a given thresholding method
to each raster index.
}
\keyword{internal}
