% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart2.R
\name{chart2}
\alias{chart2}
\title{chart2: Manual selection of points where color checker tiles are located in
the image.}
\usage{
chart2(
  pic.path,
  samp.width = 40,
  pic.format,
  interactive = F,
  tile.coords,
  xriteclassic.chart = T,
  n.color.tiles
)
}
\arguments{
\item{pic.path}{string. File path where you can find the image files.}

\item{samp.width}{numeric. Distance to establish 
the buffer around click point. This argument feeds the argument with from 
terra::buffer function operated internally by chart2. 
Default value with = 40. 
input click points, or of length equal to the number of click points.}

\item{pic.format}{character. Picture file format. It could be "jpg" for .jpg,
.JPG and .jpeg; or "tif", for .tif format.}

\item{interactive}{Logical. If set to TRUE, the function requires the user to
manually position the color tile over the image of the color card. 
If set to FALSE, you need to provide a coordinate dataframe in the 
tile.coords argument with the centroid coordinates of the color tiles. 
Default is FALSE.}

\item{tile.coords}{A dataframe containing the centroid coordinates (pixel 
column and row) of color tiles. Each row in the dataframe represents a color 
tile, and it should have columns for the X and Y coordinates.
This argument is required when the interactive parameter is set to FALSE.}

\item{xriteclassic.chart}{logical. Indicates if we are using color tiles 
from Xrite classic ColorCheker to make the color calibration. Default = TRUE. 
If FALSE (you are not using Xrite classic ColorCheker, or you want to use 
another quantity of color tiles) you need to provide the number of tiles 
you would use in the n.color.tiles parameter}

\item{n.color.tiles}{numeric. Only required if you would not use the color 
tiles from Xrite Classic ColorCheker to make the color calibration. In that 
case you need to provide the number of tiles (n) you would use in your color 
chart to make the color calibration.}
}
\value{
A spatial.polygon with 24 features one by each color tile in case 
xriteclassic.chart = T, or n features if xriteclassic.chart = F and 
n.color.tiles = n.
}
\description{
Manual selection of points where color checker tiles are located in the image.
You need to click over color tiles in the opened window where you can find 
the picture with the color chart. 
To click over the tiles you need to follow the order shown in
\href{https://raw.githubusercontent.com/MMolBus/photomoss/master/vignettes/vignette_Photomoss_workflow/chart.png}{this image}.
Preferably, click on the color tile centroids.
}
\examples{
#'
chart2(pic.path="./JPG", samp.width = 0.01, pic.format = "jpg", 
xriteclassic.chart = T, n.color.tiles)
chart2(pic.path="./JPG", samp.width = 0.01, pic.format = "jpg", 
xriteclassic.chart = F, n.color.tiles = 4) # if you only want to calibrate
with a color chart with 4 tiles

}
\author{
Manuel Molina-Bustamante
}
