% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptor.calculation.fun.R
\name{descriptor.fun}
\alias{descriptor.fun}
\title{Compute Descriptive Statistics}
\usage{
descriptor.fun(x, des)
}
\arguments{
\item{x}{A numeric vector.}

\item{des}{A character vector specifying which descriptive statistics to 
compute.
Available options: 
`"median"`, `"mean"`, `"sd"`, `"min"`, `"max"`, and `"diff.range"`.}
}
\value{
A named numeric vector containing the computed statistics.
}
\description{
This internal function computes various descriptive statistics 
for a given numeric vector based on the requested descriptors.
}
\keyword{internal}
