% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexcalculation.fun.ccsdf.R
\name{index.calc.fun}
\alias{index.calc.fun}
\title{Calculate Various Remote Sensing Indices}
\usage{
index.calc.fun(raster.mat, raster.band, index.)
}
\arguments{
\item{raster.mat}{A matrix containing raster data.}

\item{raster.band}{A raster stack containing the spectral bands.}

\item{index.}{A vector of indices to be calculated. 
It have to be one of this values "NDVI", "SR", "MSAVI", "EVI", "CI", "BI", "NORR", "NORG", "NORB",}
}
\value{
A list of raster layers corresponding to the requested indices.
}
\description{
This function computes various spectral indices from raster data.
}
\keyword{#}
\keyword{Mark}
\keyword{as}
\keyword{function}
\keyword{internal}
\keyword{the}
