% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.jpg.ccspectral.ccsdf.R
\name{raster.jpg.ccspectral}
\alias{raster.jpg.ccspectral}
\title{Process Spectral Raster Images in JPEG format (Internal)}
\usage{
raster.jpg.ccspectral(vis.photo, nir.photo, manual.mask.test, mask.photo)
}
\arguments{
\item{vis.photo}{A character string specifying the filename of the visible 
spectrum (RGB) image.}

\item{nir.photo}{A character string specifying the filename of the NIR image.}

\item{manual.mask.test}{A logical value. If TRUE, the function applies a 
binary mask (from a TIFF file).}

\item{mask.photo}{A character string specifying the filename of the binary 
mask (if manual.mask.test is TRUE).}
}
\value{
A raster object containing the combined bands: visible red, green, 
blue, NIR blue, and optionally a binary mask.
}
\description{
This function reads visible and near-infrared (NIR) spectrum images in  JPEG 
format, applies an optional binary mask, and returns a combined raster object 
with spectral bands. This function is intended for internal use only.
}
\examples{
# Example 1: Without mask
result <- .internal_raster_tiff_ccspectral("visible.tiff", "nir.tiff", FALSE, NULL)

# Example 2: With mask
result <- .internal_raster_tiff_ccspectral("visible.tiff", "nir.tiff", TRUE, "mask.tiff")

}
\keyword{#}
\keyword{Mark}
\keyword{as}
\keyword{function}
\keyword{internal}
\keyword{the}
