% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSS.IoU.calc.R
\name{TSS.IoU.calc}
\alias{TSS.IoU.calc}
\title{@title TSS and IoU Calculator}
\usage{
TSS.IoU.calc(df.result)
}
\arguments{
\item{df.result}{A data frame where each row represents a pixel classified as
0 (absence) or 1 (presence) by the classification model and the ground truth.
The data frame contains the following columns:  
  - `TP`: True Positives  
  - `FP`: False Positives  
  - `TN`: True Negatives  
  - `FN`: False Negatives}
}
\value{
A named numeric vector with two elements:
  - `"TSS"`: The computed True Skill Statistics value.
  - `"IoU"`: The computed Intersection over Union value.
}
\description{
Computes the True Skill Statistics (TSS) and Intersection over 
Union (IoU).

- True Skill Statistics (TSS) is derived from sensitivity and specificity, 
  calculated using presence and absence records, and is used for assessing 
  model performance.

- Intersection over Union (IoU) is commonly used in image segmentation 
  performance assessment by comparing the ground truth polygon with the 
  predicted polygon.
}
\keyword{internal}
